little_genie/little_genie/requirements.txt
Django==4.0.3
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-import-export==2.8.0
django-registration-redux==2.10
django-tinymce==3.4.0
django-versatileimagefield==2.2
djangorestframework==3.13.1
drf-yasg==1.20.0
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
requests==2.27.1
